Use Ansible tag to skip tasks that would remove root SSH access
We only skip these tasks if we're using the initial root SSH access,
as removing it would not allow us to complete the following tasks.

This will ensure that the tasks still get run later, when using Ansible
with sudo instead of direct root SSH access. This is important for
documentation purposes, and to ensure that the settings remain as we intend.